more test cases for sha3 and big number
